upload_231112_2

+++ tatyamyam/src/main/java/com/example/tatyamyam/repository/HealthRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.tatyamyam.entity.Health;
 
-public interface HealthRepository extends JpaRepository<Health, Integer> {
+public interface HealthRepository extends JpaRepository<Health, Long> {
 
 	Health findByUserIdAndDate(Long userID, LocalDate today);
